Changed Text Engine Roll
To make text roll, you now need to set TEXT_ROLL as an argument to fillWithSentence (default no roll) or set rollText to true. This allowed for more stability in rolling text by keeping track of the last sentence to use roll. If the same sentence is used multiple times in a row, it will only roll the first time. Now the player can press the B button to quickly finish a sentence if they do not want to wait without the sentence repeating.

Additionally, now the player cannot waste a turn when trying to cast magic when the player does not have enough MP even if the player is slower than the enemy.